LayoutTests/platform/wpe/mathml/non-core/lengths-3-expected.txt


FAIL Units assert_equals: ex expected 500 but got 499.984375
FAIL Trimming of space assert_equals: ex expected 500 but got 499.984375
PASS Non-negative numbers
PASS Non-positive numbers
